Removal: Remove favicon.ico
The commit history for this repo is going to look terrible. I deleted this file because the error persists, because like with the background image, I need to tell the site to not look for its relative path, but for its raw link on the repo itself. Since I'm not sure how to do this yet and need to go to sleep today, I'm removing it entirely, and will redo once I've made that better icon I mentioned in the last commit.